HAI-2576 Refactor useHaittaIndexes and useHaittaIndexSummary hooks
